refactor:coloca repository e service dentro das pastas workout e renomeia namespaces.
